Fixing project structure.
