+++ src/services/databaseService.js
@@ -55,3 +55,5 @@ async function saveBooksToMongoDB(books) {
         }
     }
 }
+
+module.exports =  saveBooksToMongoDB;
